+++ react_training/src/Day4/CSS/Task4.css
@@ -1,0 +1,10 @@
+.mainDiv{
+    gap: 7px;
+    justify-content: center;
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: column;
+}
+button{
+    width: 100px;
+}
