+++ react_training/src/Day4/Components/Task12/Task12.js
@@ -1,0 +1,4 @@
+import getComparator from "./getComparator";
+import EnhancedTableToolbar from "./EnhancedTableToolbar";
+import EnhancedTableHead from "./EnhancedTableHead";
+export { getComparator, EnhancedTableToolbar, EnhancedTableHead };
