+++ react_training/src/Day4/Day4.js
@@ -1,37 +1,52 @@
-import React from 'react'
-import Task13 from './Task13/Task13'
-import Task14 from './Task14/Task14'
-// import Task1 from './Task1/Task1'
-// import Task2 from './Task2/Task2'
-// import Task4 from './Task4/Task4'
-// import Task3 from './Task3/Task3'
-// import Task5 from './Task5/Task5'
-// import Task6 from './Task6/Task6'
-// import Task7 from './Task7/Task7'
-// import Task8 from './Task8/Task8'
-// import Task9 from './Task9/Task9'
-// import FormValidation from './Task10/Task10'
-// import Task11 from './Task11/Task11'
-// import Task12 from './Task12/Task12'
+import React from "react";
+import {
+  Task1,
+  Task10,
+  Task11,
+  Task12,
+  Task13,
+  Task14,
+  Task2,
+  Task3,
+  Task4,
+  Task5,
+  Task6,
+  Task7,
+  Task8,
+  Task9,
+} from "./index";
 
 const Day4 = () => {
   return (
     <div>
-      {/* <Task1 />
+      <Task1 />
+      <hr />
       <Task2 />
+      <hr />
       <Task3 />
+      <hr />
       <Task4 />
+      <hr />
       <Task5 />
+      <hr />
       <Task6 />
+      <hr />
       <Task7 />
+      <hr />
       <Task8 />
+      <hr />
       <Task9 />
-      <FormValidation />
+      <hr />
+      <Task10 />
+      <hr />
       <Task11 />
-      <Task12 /> */}
-      {/* <Task13 /> */}
+      <hr />
+      <Task12 />
+      <hr />
+      <Task13 />
+      <hr />
       <Task14 />
     </div>
-  )
-}
-export default Day4
+  );
+};
+export default Day4;
