+++ react_training/src/Day4/Task1/Task1.jsx
@@ -2,6 +2,7 @@
 Implement it as a controlled component where the input value is controlled by the component's state. 
 When the user types into the input field, the component's state should update accordingly. */
 import React, { useState } from "react";
+import {Text,type_here_placeholder} from '../constants.js'
 
 const Task1 = () => {
   const [inputValue, setInputValue] = useState("");
@@ -9,10 +10,10 @@ const Task1 = () => {
     <div>
       <h1>Task1</h1>
       <input
-        type="text"
+        type={Text}
         value={inputValue}
         onChange={(e) => setInputValue(e.target.value)}
-        placeholder="Type Here..."
+        placeholder={type_here_placeholder}
       />
       <p>Input value: {inputValue}</p>
     </div>
