+++ react_training/src/Day4/constants.js
@@ -1,0 +1,37 @@
+const Text = "text";
+const type_here_placeholder = "type here......";
+const Checkbox = "checkbox";
+const Name = "name";
+const Email = "email";
+const cap_Email = "Email";
+const Number = "number";
+const Age = "age";
+const Radio = "radio";
+const Gender = "gender";
+const Male = "male";
+const Female = "female";
+const Agree = "agree";
+const Submit = "submit";
+const cap_Submit = "Submit";
+const Password = "password";
+const cap_password = "Password";
+
+export {
+  cap_Email,
+  Password,
+  cap_password,
+  cap_Submit,
+  Agree,
+  Submit,
+  Text,
+  type_here_placeholder,
+  Checkbox,
+  Name,
+  Email,
+  Number,
+  Age,
+  Radio,
+  Gender,
+  Male,
+  Female,
+};
